UPSTREAM: Smack: Privilege check on key operations
Smack: Privilege check on key operations

Operations on key objects are subjected to Smack policy
even if the process is privileged. This is inconsistent
with the general behavior of Smack and may cause issues
with authentication by privileged daemons. This patch
allows processes with CAP_MAC_OVERRIDE to access keys
even if the Smack rules indicate otherwise.

Reported-by: Jose Bollo <jobol@nonadev.net>
Change-Id: I0803b899e536695f0f31e6c443a3e4e7f22d082f
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Vayruz Rafli <mikhailovnakowalski@gmail.com>