BACKPORT: vfs: do bulk POLL* -> EPOLL* replacement
This is the mindless scripted replacement of kernel use of POLL*
variables as described by Al, done by this script:

    for V in IN OUT PRI ERR RDNORM RDBAND WRNORM WRBAND HUP RDHUP NVAL MSG; do
        L=`git grep -l -w POLL$V | grep -v '^t' | grep -v /um/ | grep -v '^sa' | grep -v '/poll.h$'|grep -v '^D'`
        for f in $L; do sed -i "-es/^\([^\"]*\)\(\<POLL$V\>\)/\\1E\\2/" $f; done
    done

with de-mangling cleanups yet to come.

NOTE! On almost all architectures, the EPOLL* constants have the same
values as the POLL* constants do.  But they keyword here is "almost".
For various bad reasons they aren't the same, and epoll() doesn't
actually work quite correctly in some cases due to this on Sparc et al.

The next patch from Al will sort out the final differences, and we
should be all done.

[Linux4: Only change drivers/media/rc/lirc_dev.c]

Scripted-by: Al Viro <viro@zeniv.linux.org.uk>
Change-Id: I154ca15a54f61707547ad97b22b0af3dd43ad4a4
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Vayruz Rafli <mikhailovnakowalski@gmail.com>